Adjust tests for asyncSubject
