+++ package.json
@@ -41,6 +41,7 @@
 	},
 	"prettier": {
 		"singleQuote": true,
-		"tabWidth": 2
+		"tabWidth": 2,
+        "arrowParens": "avoid"
 	}
 }
